requirements.txt
pydantic==2.11.7
pydantic-settings==2.10.1
pillow==11.3.0
opencv-python==4.12.0.88
rembg==2.0.67
fastapi==0.116.1
uvicorn==0.35.0
ecs-logging==2.2.0
onnxruntime==1.22.1
python-multipart==0.0.20